requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
jupyter==1.0.0
pillow==10.0.0
scikit-learn==1.3.0
